feat : update FirebaseConfig
