+++ src/main/java/com/bbf/bebefriends/global/config/FirebaseConfig.java
@@ -3,42 +3,33 @@ package com.bbf.bebefriends.global.config;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
-import lombok.RequiredArgsConstructor;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.ResourceLoader;
 
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
 @Configuration
-@RequiredArgsConstructor
+
 public class FirebaseConfig {
-    private final ResourceLoader resourceLoader;
 
-    @Value("${firebase.bucket-name}")
-    private String bucketName;          // firebase 버킷 이름
+    @Value("${firebase.config-path}")
+    private String firebaseConfigPath;
 
-    @Value("${firebase.credentials.path}")
-    private String credentialsPath;     // firebase 인증 json
+    @PostConstruct
+    public void initialize() {
+        try (InputStream serviceAccount = new FileInputStream(firebaseConfigPath)) {
+            FirebaseOptions options = FirebaseOptions.builder()
+                    .setCredentials(GoogleCredentials.fromStream(serviceAccount))
+                    .build();
 
-    @Bean
-    public FirebaseApp firebaseApp() throws IOException {
-        Resource resource = resourceLoader.getResource(credentialsPath);
-        InputStream serviceAccount = resource.getInputStream();
-
-        FirebaseOptions options = FirebaseOptions.builder()
-                .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                .setStorageBucket(bucketName)
-                .build();
-
-        if (FirebaseApp.getApps().isEmpty()) {
-            return FirebaseApp.initializeApp(options);
-        } else {
-            return FirebaseApp.getInstance();
+            if (FirebaseApp.getApps().isEmpty()) {
+                FirebaseApp.initializeApp(options);
+            }
+        } catch (IOException e) {
+            throw new IllegalStateException("Firebase 초기화 실패", e);
         }
     }
 }
