--- src/main/java/com/bbf/bebefriends/member/config/FirebaseConfig.java
@@ -1,35 +1,0 @@
-package com.bbf.bebefriends.member.config;
-
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import jakarta.annotation.PostConstruct;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-
-@Configuration
-public class FirebaseConfig {
-
-    @Value("${firebase.config-path}")
-    private String firebaseConfigPath;
-
-    @PostConstruct
-    public void initialize() {
-        try (InputStream serviceAccount = new FileInputStream(firebaseConfigPath)) {
-            FirebaseOptions options = FirebaseOptions.builder()
-                    .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                    .build();
-
-            if (FirebaseApp.getApps().isEmpty()) {
-                FirebaseApp.initializeApp(options);
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException("Firebase 초기화 실패", e);
-        }
-    }
-}
